refactor(markers): rename `useMarkers` to `use-markers` for consistency with file naming conventions
